new desktop layout for controls
