+++ agency-dashboard/src/DashboardView.tsx
@@ -16,32 +16,25 @@
 // =============================================================================
 
 import { ReactComponent as DownloadIcon } from "@justice-counts/common/assets/download-icon.svg";
-import { ReactComponent as GridIcon } from "@justice-counts/common/assets/grid-icon.svg";
 import { ReactComponent as InfoIcon } from "@justice-counts/common/assets/info-icon.svg";
 import { ReactComponent as LeftArrowIcon } from "@justice-counts/common/assets/left-arrow-icon.svg";
 import { ReactComponent as ShareIcon } from "@justice-counts/common/assets/share-icon.svg";
 import { DatapointsView } from "@justice-counts/common/components/DataViz/DatapointsView";
-import { ExtendedDropdownMenuItem } from "@justice-counts/common/components/DataViz/DatapointsView.styles";
-import { DropdownMenu, DropdownToggle } from "@recidiviz/design-system";
 import { observer } from "mobx-react-lite";
 import React, { useEffect } from "react";
 import { useLocation, useNavigate, useParams } from "react-router-dom";
 
 import {
-  AllMetricsButtonContainer as SelectMetricButtonContainer,
-  AllMetricsButtonText,
   Container,
-  ExtendedDropdown,
   LeftPanel,
   LeftPanelBackButtonContainer,
   MetricOverviewActionButtonContainer,
   MetricOverviewActionButtonText,
   MetricOverviewActionsContainer,
   MetricOverviewContent,
-  MetricOverviewTitle,
   MetricTitle,
   RightPanel,
-  RightPanelTopContainer,
+  // RightPanelTopContainer,
 } from "./DashboardView.styles";
 import { HeaderBar } from "./Header/HeaderBar";
 import { useStore } from "./stores";
@@ -73,36 +66,6 @@ const MetricOverviewActionShareButton = () => (
     <ShareIcon />
     <MetricOverviewActionButtonText>Share</MetricOverviewActionButtonText>
   </MetricOverviewActionButtonContainer>
-);
-
-const SelectMetricButton = () => (
-  <SelectMetricButtonContainer>
-    <GridIcon />
-    <AllMetricsButtonText />
-  </SelectMetricButtonContainer>
-);
-
-const SelectMetricButtonDropdown: React.FC<{
-  onSelect: (metricKey: string) => void;
-  options: string[];
-}> = ({ onSelect, options }) => (
-  <ExtendedDropdown>
-    <DropdownToggle>
-      <SelectMetricButton />
-    </DropdownToggle>
-    <DropdownMenu>
-      {options.map((value) => (
-        <ExtendedDropdownMenuItem
-          key={value}
-          onClick={() => {
-            onSelect(value);
-          }}
-        >
-          {value}
-        </ExtendedDropdownMenuItem>
-      ))}
-    </DropdownMenu>
-  </ExtendedDropdown>
 );
 
 const DashboardView = () => {
@@ -155,7 +118,6 @@ const DashboardView = () => {
         <MetricTitle>
           {datapointsStore.metricKeyToDisplayName[metricKey] || metricKey}
         </MetricTitle>
-        <MetricOverviewTitle />
         <MetricOverviewContent>
           Measures the number of individuals with at least one parole violation
           during the reporting period.
@@ -167,20 +129,18 @@ const DashboardView = () => {
         </MetricOverviewActionsContainer>
       </LeftPanel>
       <RightPanel>
-        <RightPanelTopContainer>
-          <SelectMetricButtonDropdown
-            onSelect={(metric) =>
-              navigate(`/agency/${agencyId}/dashboard?metric=${metric}`)
-            }
-            options={metricNames}
-          />
-        </RightPanelTopContainer>
+        {/* <RightPanelTopContainer>
+        </RightPanelTopContainer> */}
         <DatapointsView
           datapointsGroupedByAggregateAndDisaggregations={
             datapointsStore.datapointsByMetric[metricKey]
           }
           dimensionNamesByDisaggregation={
             datapointsStore.dimensionNamesByMetricAndDisaggregation[metricKey]
+          }
+          metricNames={metricNames}
+          onMetricsSelect={(metric) =>
+            navigate(`/agency/${agencyId}/dashboard?metric=${metric}`)
           }
         />
       </RightPanel>
